Get request for mentors moved to Redux Store.
